*sigh* Heroku forces my app to put Rails at toplevel
Looks like foreman can take an optional root, but you have to do this with a flag when invoking it.
I can't figure out where it gets invoked, or if all these things are done by something else.
Maybe if I dig through the heroku buildpack, I could find it, but based on past experiences,
this is going to ultimately be futile.